add vfsgnj test

+++ my_tests/src/main.cpp
@@ -227,7 +227,7 @@ int main(int argc, char *argv[])
     // vfsgnj
     if (1) {
         printf("\n*****vmflt_vv_f64 test*****\n");
-        test_2src_op<int64_t, double, double>(vfsgnj_vv_f64_ref, vfsgnj_vv_f64_vec);
+        test_2src_op<double, double, double>(vfsgnj_vv_f64_ref, vfsgnj_vv_f64_vec);
     }
     return 0;
 }
